+++ fdrh-impl/src/main/java/com/dwarfeng/fdrh/impl/dao/processor/PersistenceValueHandle.java
@@ -11,11 +11,21 @@ import org.springframework.stereotype.Component;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Objects;
 
+@SuppressWarnings("DuplicatedCode")
 @Component
 public class PersistenceValueHandle implements BaseHandle<LongIdKey, PersistenceValue>,
         PresetLookupHandle<PersistenceValue> {
+
+    private static final Map<String, QueryInfo.Ordering> ORDERING_MAP;
+
+    static {
+        ORDERING_MAP = new HashMap<>();
+        ORDERING_MAP.put("happened_date", QueryInfo.Ordering.ASC);
+    }
 
     @Override
     public PersistenceValue newInstance() {
@@ -93,7 +103,7 @@ public class PersistenceValueHandle implements BaseHandle<LongIdKey, Persistence
         Object[] parameters;
         whereClause = "happened_date >= ? AND happened_date < ?";
         parameters = new Object[]{args[0], args[1]};
-        return new QueryInfo(whereClause, null, parameters);
+        return new QueryInfo(whereClause, ORDERING_MAP, parameters);
     }
 
     private QueryInfo presetChildForPoint(Object[] args) {
@@ -107,7 +117,7 @@ public class PersistenceValueHandle implements BaseHandle<LongIdKey, Persistence
             whereClause = "point_id=?";
             parameters = new Object[]{key.getLongId()};
         }
-        return new QueryInfo(whereClause, null, parameters);
+        return new QueryInfo(whereClause, ORDERING_MAP, parameters);
     }
 
     private QueryInfo presetChildForPointBetween(Object[] args) {
@@ -121,6 +131,6 @@ public class PersistenceValueHandle implements BaseHandle<LongIdKey, Persistence
             whereClause = "point_id = ? AND happened_date >= ? AND happened_date < ?";
             parameters = new Object[]{key.getLongId(), args[1], args[2]};
         }
-        return new QueryInfo(whereClause, null, parameters);
+        return new QueryInfo(whereClause, ORDERING_MAP, parameters);
     }
 }